Check if project exists, before showing anything
